OpenCV set detection working but too many sets are often found

+++ eink-common/src/main/java/radomik/com/github/dto/BarInfoDto.java
@@ -5,22 +5,22 @@ import java.io.Serializable;
 public class BarInfoDto implements Serializable {
     private static final long serialVersionUID = -555492705961890304L;
 
-    private int y0;
-    private int height;
+    private short y0;
+    private short height;
 
-    public int getY0() {
+    public short getY0() {
         return y0;
     }
 
-    public void setY0(int y0) {
+    public void setY0(short y0) {
         this.y0 = y0;
     }
 
-    public int getHeight() {
+    public short getHeight() {
         return height;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(short height) {
         this.height = height;
     }
