+++ eink-common/src/main/java/radomik/com/github/dto/PointDto.java
@@ -5,22 +5,22 @@ import java.io.Serializable;
 public class PointDto implements Serializable {
     private static final long serialVersionUID = 1384899799243658211L;
 
-    private int x;
-    private int y;
+    private short x;
+    private short y;
 
-    public int getX() {
+    public short getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(short x) {
         this.x = x;
     }
 
-    public int getY() {
+    public short getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(short y) {
         this.y = y;
     }
